requirements.txt
aiogram>=3.7.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
redis>=5.0.0
aiohttp>=3.10.0
pydantic>=2.9.0
jinja2>=3.1.0
python-multipart>=0.0.9
cachetools>=5.5.0
sqlalchemy>=2.0.0
alembic>=1.13.0
cryptography>=43.0.0